281: create look/setup.js
