+++ src/look/locations.js
@@ -1,12 +1,9 @@
 import _ from 'lodash';
 
 import { PPTL } from '../global/constants';
-import { assert } from '../global/utils';
-import { tileHasName, tileIsOneOf } from './tileInfo';
+import { getLocations } from './setup';
+import { tileHasName } from './tileInfo';
 import { isOnMyTeam } from './gameState';
-
-
-const locations = {};
 
 
 /**
@@ -14,7 +11,7 @@ const locations = {};
  * @returns {{xt: number, yt: number}} a tile which is at the center of mass of all occurrences of
  *   tileName in the tagpro map
  */
-function centerOfMass(tileName) {
+export function centerOfMass(tileName) {
   let xSum = 0;
   let ySum = 0;
   let count = 0;
@@ -33,31 +30,6 @@ function centerOfMass(tileName) {
 
 
 /**
- * Parses each tile of the TagPro map. If it discovers a flag, add its location to the locations
- *   object. Also defines the endzone positions for red and blue teams.
- */
-export function initLocations() {
-  assert(tagpro.map, 'tagpro.map is undefined');
-  locations.BLUE_ENDZONE = centerOfMass('BLUE_ENDZONE');
-  locations.RED_ENDZONE = centerOfMass('RED_ENDZONE');
-  for (let xt = 0, xl = tagpro.map.length; xt < xl; xt++) {
-    for (let yt = 0, yl = tagpro.map[0].length; yt < yl; yt++) {
-      if (tileIsOneOf(tagpro.map[xt][yt], ['RED_FLAG', 'RED_FLAG_TAKEN'])) {
-        locations.RED_FLAG = { xt, yt };
-        locations.RED_FLAG_TAKEN = { xt, yt };
-      } else if (tileIsOneOf(tagpro.map[xt][yt], ['BLUE_FLAG', 'BLUE_FLAG_TAKEN'])) {
-        locations.BLUE_FLAG = { xt, yt };
-        locations.BLUE_FLAG_TAKEN = { xt, yt };
-      } else if (tileIsOneOf(tagpro.map[xt][yt], ['YELLOW_FLAG', 'YELLOW_FLAG_TAKEN'])) {
-        locations.YELLOW_FLAG = { xt, yt };
-        locations.YELLOW_FLAG_TAKEN = { xt, yt };
-      }
-    }
-  }
-}
-
-
-/**
  * Returns the position xp and yp (in pixels) of the center of one of the specified tile
  *   types. Assumes that the potential location of the tile has been stored by calling
  *   initLocations(). Runtime: O(1)
@@ -65,6 +37,7 @@ export function initLocations() {
  *   or an array of such numbers
  */
 export function findCachedTile(tileNames) {
+  const locations = getLocations();
   const tileNameArray = [].concat(tileNames);
   for (let i = 0; i < tileNameArray.length; i++) {
     const name = tileNameArray[i];
